Do not use simsimd in linux/arm64 (#2455)
manylinux-cross, the docker image used to cross-compile for arm, uses gcc 7.5.0. However, simsimd, requires the arm SVE extension which was added to GCC in version 8. Thus, it doesn't compile correctly in the image and produces broken bindings.

The solution used here is to compile the non-simd version of the similarity functions for this specific case, at least until a new manylinux-cross version is released with newer GCC.