Reindex all questions when adding a new one (#2526)
## Always send all splits to the index each time
This makes it so we send the entire conversation field to the index when adding a new question. This way the granularity of the index is the field and it doesn't have to know about splits and could make our code simpler.

## Remove splits from `paragraphs_to_delete` and `sentences_to_delete`
We are always sending all splits, so we can reindex the entire field

## Remove per-paragraph deletion from `paragraphs_to_delete`
This was started a few months ago and the changes are already propagated everywhere

## Remove vectors on field deletion
When deleting a field, `delete_metadata()` was called which deleted only from paragraph index, not vectors index.