Nidx vector cleanup (#2569)
* Remove nidx_vector service layer

Remove the service layer from the vector index:
- The reader had logic parsing the search request, this has been moved
  to the data point provider reader, including tests.
- The writer has been deleted. The set_resource method has been
  extracted to indexer.rs as index_resource.

* Remove vector index state and writer

Remove the state files shared by reader and writer and instead pass the
metadata in function calls. The data point writer has been completely
removed since its main function was updating that state. The reader
takes the state it needs from the constructor instead of reading it from
disk.

* Remove nidx_vector pins

We used to have pins as a means to lock a segment while being opened.
This is not needed anymore, so removing the pin concept altogether.

* Remove journal from vectors index

The Journal that was serialized to disk is removed since the metadata
will now be handled externally. We now return metadata when a segment is
created and require the same metadata to be passed when opening it.

Pending topic: a better way to handle deletions

* Remove unneeded serialization dependencies